Move dispatch_aborted_exception to its own file, add default ctor
Adding an exception explanation behavior to `dispatch_aborted_exception` is nice, but it still must be possible to default-construct this type.  Also provide it its own header.
